refactor(obd2): extract TripLiveReading + distance source constants (Refs #563)
Split TripLiveReading DTO and distance-source / virtual-odometer cap
constants out of trip_recording_controller.dart (1117 -> 1063 LOC) into
two new cohesive files. Makes the constants public (kDistanceSourceReal,
kDistanceSourceVirtual, kVirtualOdometerSampleCap) so callers that
inspect a finalised TripSummary can compare against named constants
instead of magic strings.

Controller re-exports both modules so existing callers (providers,
widget tests) keep working without import churn. This is phase 1 of
the #563 controller split — further phases can pull out the scheduler
wiring, drop-detection state machine, and fuel-rate derivation.
